requirements.txt
telethon>=1.36
pyyaml>=6.0
tqdm>=4.66
aiogram>=3.15
anthropic>=0.42
openai>=1.50
python-dotenv>=1.0
